Adição de funcionalidades no Combate

+++ simulador-rpg/src/Entidades/Inimigos/Enemy.cs
@@ -3,7 +3,7 @@ namespace Entidades.Inimigos
     public class Enemy
     {
         public string name { get; set; }
-        public float hp { get; set; }
-        public float ataque { get; set; }
+        public double hp { get; set; }
+        public double ataque { get; set; }
     }
 }
